pyproject.toml
[project]
name = "pal_module"
version = "0.3.0"
description = "Software for automatting a Vention Rails"
authors = [
    {name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"},
    {name = "Rafael Vescovi", email="ravescovi@anl.gov"},
    {name = "Doga Ozgulbas", email="dozgulbas@anl.gov"},
    {name = "Abe Stroka", email="astroka@anl.gov"},
    {name = "Kyle Hippe", email = "khippe@anl.gov"},
    {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
]
dependencies = [
    "madsci.node_modules",
    "fastapi>=0.103.2",
    "uvicorn>=0.21.1",
    "pytest",
    "numpy",
    "paho-mqtt",

]

requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/vention_rail_module"

######################
# Build Info + Tools #
######################
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

#####################
# Development Tools #
#####################

[tool.ruff]
# https://docs.astral.sh/ruff/configuration/

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "docs",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    # "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    # "SIM",
    # isort
    "I",
    # Warning
    "W",
    # pydocstyle
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    # ruff
    # "RUF"
]
ignore = [
    "E501", # Line too long
    "B006", # Do not use mutable data structures for argument defaults
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/customize.html
addopts = "-x"
junit_family="xunit1"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pottery.exceptions.InefficientAccessWarning",
]
markers = [
    "hardware: marks test as requiring hardware (deselect with '-m \"not hardware\"')",
]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
show_error_codes = true
check_untyped_defs = true
follow_imports = "normal"
strict_optional = true
plugins = ["pydantic.mypy"]
strict = true
disallow_untyped_defs = true
implicit_reexport = true
